backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-google-genai>=0.0.3
pandas==2.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sse-starlette==1.8.2
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
reportlab==4.0.7

